Moved to more hierarchical structure. Changed relative imports to absolute imports, which is imposed in Golang
