+++ indexBuilders/builder.go
@@ -1,7 +1,7 @@
 package indexBuilders
 
 import (
-	"../indexes"
+	"github.com/rmulton/riw_project/indexes"
 )
 
 type IndexBuilder interface {
