+++ indexBuilders/inMemory/InMemoryBuilder_test.go
@@ -1,18 +1,16 @@
-package indexBuilders
+package inMemory
 
 import (
-	"os"
 	"testing"
 	"sync"
 	"reflect"
-	"../indexes"
-	"../utils"
+	"github.com/rmulton/riw_project/indexes"
 )
 
-var onDiskWaitGroup sync.WaitGroup
-var onDiskReadingChannel = make(indexes.ReadingChannel)
+var testInMemWaitGroup sync.WaitGroup
+var testInMemReadingChannel = make(chan indexes.Document)
 
-var someDocuments = []indexes.Document {
+var someDocuments2 = []indexes.Document {
 	indexes.Document {
 		ID: 0,
 		Path: "./mock_path/test.test",
@@ -90,6 +88,30 @@ var someDocuments = []indexes.Document {
 			"blabla",
 			"blabla",
 			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
 
 			"bleble",
 			"bleble",
@@ -114,12 +136,12 @@ var someDocuments = []indexes.Document {
 	},
 }
 
-var expectedPostingLists = map[string]indexes.PostingList {
+var expectedPostingLists2 = map[string]indexes.PostingList {
 	"blabla": indexes.PostingList {
 		0: 1,
 		12: 2,
 		27: 1,
-		3674: 12,
+		3674: 36,
 	},
 	"bleble": indexes.PostingList {
 		0: 1,
@@ -165,7 +187,7 @@ var expectedPostingLists = map[string]indexes.PostingList {
 	},
 }
 
-var expectedDocIDToFilePath = map[int]string {
+var expectedDocIDToFilePath2 = map[int]string {
 	0: "./mock_path/test.test",
 	12: "ssljfsd",
 	64: "sdlkajfsdflkdfjd",
@@ -174,27 +196,25 @@ var expectedDocIDToFilePath = map[int]string {
 	3674: "djdsl",
 }
 
-func TestBuildOnDisk(t *testing.T) {
-	utils.ClearOrCreatePersistedIndex("./saved")
-	var builder = NewOnDiskBuilder(3, "./saved", onDiskReadingChannel, 2, &onDiskWaitGroup)
+func TestBuildInMemory(t *testing.T) {
+	var builder = NewInMemoryBuilder(testInMemReadingChannel, 2, &testInMemWaitGroup)
 	builder.parentWaitGroup.Add(1)
 	go builder.Build()
-	for _, doc := range someDocuments {
+	for _, doc := range someDocuments2 {
 		builder.readingChannel <- doc
 	}
 	close(builder.readingChannel)
-	builder.parentWaitGroup.Wait()
 	// NB: the tf-idf functionality is tested in ./indexes. Here we rely on it and keep
 	// scores as integers for clarity
-	for _, postingList := range expectedPostingLists {
-		postingList.TfIdf(len(someDocuments))
+	builder.parentWaitGroup.Wait()
+	for _, postingList := range expectedPostingLists2 {
+		postingList.TfIdf(len(someDocuments2))
 	}
 	index := builder.GetIndex()
 	// For terms that are in both indexes
-	for term, expectedPostingList := range expectedPostingLists {
+	for term, expectedPostingList := range expectedPostingLists2 {
 		postingList := index.GetPostingListsForTerms([]string{term})[term]
 		if !reflect.DeepEqual(postingList, expectedPostingList) {
-			// t.Errorf("\nFor %s:\n   - Should be %v\n   - Not %v\n", term, expectedPostingList, postingList)
 			for docID, expectedScore := range expectedPostingList {
 				score := postingList[docID]
 				if score != expectedScore {
@@ -203,6 +223,8 @@ func TestBuildOnDisk(t *testing.T) {
 			}
 		}
 	}
-	// Clear the folder after
-	os.RemoveAll("./saved")
+	// Check the docID to path map
+	if !reflect.DeepEqual(index.GetDocIDToPath(), expectedDocIDToFilePath2) {
+		t.Errorf("DocID to path is not correct")
+	}
 }
