+++ indexBuilders/onDisk/OnDiskBuilder_test.go
@@ -1,16 +1,18 @@
-package indexBuilders
+package onDisk
 
 import (
+	"os"
 	"testing"
 	"sync"
 	"reflect"
-	"../indexes"
+	"github.com/rmulton/riw_project/indexes"
+	"github.com/rmulton/riw_project/utils"
 )
 
-var testInMemWaitGroup sync.WaitGroup
-var testInMemReadingChannel = make(chan indexes.Document)
+var onDiskWaitGroup sync.WaitGroup
+var onDiskReadingChannel = make(indexes.ReadingChannel)
 
-var testInMemSomeDocuments = []indexes.Document {
+var someDocuments = []indexes.Document {
 	indexes.Document {
 		ID: 0,
 		Path: "./mock_path/test.test",
@@ -65,17 +67,88 @@ var testInMemSomeDocuments = []indexes.Document {
 			"toto",
 		},
 	},
+	indexes.Document {
+		ID: 3674,
+		Path: "djdsl",
+		NormalizedTokens: []string {
+
+			"aaa",
+			"aaa",
+			"aaa",
+			"aaa",
+			"aaa",
+
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+			"blabla",
+
+			"bleble",
+			"bleble",
+
+			"bebebe",
+			"bebebe",
+			"bebebe",
+			"bebebe",
+			"bebebe",
+			"bebebe",
+
+			"toto",
+			"toto",
+			"toto",
+			"toto",
+			"toto",
+			"toto",
+			"toto",
+			"toto",
+			"toto",
+		},
+	},
 }
 
-var expectedInMemPostingLists = map[string]indexes.PostingList {
+var expectedPostingLists = map[string]indexes.PostingList {
 	"blabla": indexes.PostingList {
 		0: 1,
 		12: 2,
 		27: 1,
+		3674: 36,
 	},
 	"bleble": indexes.PostingList {
 		0: 1,
 		64: 1,
+		3674: 2,
 	},
 	"blublu": indexes.PostingList {
 		0: 1,
@@ -83,6 +156,7 @@ var expectedInMemPostingLists = map[string]indexes.PostingList {
 	"aaa": indexes.PostingList {
 		12: 1,
 		64: 3,
+		3674: 5,
 	},
 	"bbb": indexes.PostingList {
 		12: 1,
@@ -107,39 +181,44 @@ var expectedInMemPostingLists = map[string]indexes.PostingList {
 	},
 	"bebebe": indexes.PostingList {
 		324: 1,
+		3674: 6,
 	},
 	"toto": indexes.PostingList {
 		324: 2,
+		3674: 9,
 	},
 }
 
-var expectedInMemDocIDToFilePath = map[int]string {
+var expectedDocIDToFilePath = map[int]string {
 	0: "./mock_path/test.test",
 	12: "ssljfsd",
 	64: "sdlkajfsdflkdfjd",
 	27: "dlfkdsl",
 	324: "dflkjdsl",
+	3674: "djdsl",
 }
 
-func TestBuildInMemory(t *testing.T) {
-	var builder = NewInMemoryBuilder(testInMemReadingChannel, 2, &testInMemWaitGroup)
+func TestBuildOnDisk(t *testing.T) {
+	utils.ClearOrCreatePersistedIndex("./saved")
+	var builder = NewOnDiskBuilder(3, onDiskReadingChannel, 2, &onDiskWaitGroup)
 	builder.parentWaitGroup.Add(1)
 	go builder.Build()
-	for _, doc := range testInMemSomeDocuments {
+	for _, doc := range someDocuments {
 		builder.readingChannel <- doc
 	}
 	close(builder.readingChannel)
+	builder.parentWaitGroup.Wait()
 	// NB: the tf-idf functionality is tested in ./indexes. Here we rely on it and keep
 	// scores as integers for clarity
-	builder.parentWaitGroup.Wait()
-	for _, postingList := range expectedInMemPostingLists {
+	for _, postingList := range expectedPostingLists {
 		postingList.TfIdf(len(someDocuments))
 	}
 	index := builder.GetIndex()
 	// For terms that are in both indexes
-	for term, expectedPostingList := range expectedInMemPostingLists {
+	for term, expectedPostingList := range expectedPostingLists {
 		postingList := index.GetPostingListsForTerms([]string{term})[term]
 		if !reflect.DeepEqual(postingList, expectedPostingList) {
+			// t.Errorf("\nFor %s:\n   - Should be %v\n   - Not %v\n", term, expectedPostingList, postingList)
 			for docID, expectedScore := range expectedPostingList {
 				score := postingList[docID]
 				if score != expectedScore {
@@ -148,4 +227,10 @@ func TestBuildInMemory(t *testing.T) {
 			}
 		}
 	}
+	// Check the docID to path map
+	if !reflect.DeepEqual(index.GetDocIDToPath(), expectedDocIDToFilePath) {
+		t.Errorf("DocID to path is not correct")
+	}
+	// Clear the folder after
+	os.RemoveAll("./saved")
 }
