+++ indexes/building/BuildingIndex.go
@@ -1,0 +1,7 @@
+package indexes
+
+type BuildingIndex interface {
+	AddDocToTerm(int, string)
+	AddDocToIndex(int, string)
+	GetDocCounter() int
+}
