+++ readers/cacm/cacm.go
@@ -7,9 +7,9 @@ import (
 	"sync"
 	"log"
 	"strings"
-	"../normalizers"
-	"../utils"
-	"../indexes"
+	"github.com/rmulton/riw_project/normalizers"
+	"github.com/rmulton/riw_project/utils"
+	"github.com/rmulton/riw_project/indexes"
 )
 
 type CACMReader struct {
